+++ crates/compile/Cargo.toml
@@ -23,10 +23,11 @@ anyhow = { workspace = true }
 # cfg-if = { workspace = true }
 # gimli = { workspace = true }
 object = { workspace = true }
-# serde = "1.0.188"
-# serde_derive = "1.0.188"
+serde = "1.0.188"
+serde_derive = "1.0.188"
 # addr2line = { version = "0.21.0", default-features = false, optional = true }
-# bincode = "1.2.1"
+bincode = "1.2.1"
+wasmparser = { workspace = true }
 # rustc-demangle = { version = "0.1.16", optional = true }
 # cpp_demangle = { version = "0.3.2", optional = true }
 # log = { workspace = true }
